src/constraintsuite/utils.py
"""
Utility functions for ConstraintSuite.

This module provides common utilities for:
- Configuration loading and validation
- JSONL I/O operations
- Logging setup
- Random seed management
"""

from pathlib import Path
from typing import Any, Iterator
import json
import logging
import random

import yaml
import numpy as np


def load_config(config_path: str | Path) -> dict[str, Any]:
    """
    Load and validate a YAML configuration file.

    Args:
        config_path: Path to the YAML configuration file.

    Returns:
        Dictionary containing configuration parameters.

    Raises:
        FileNotFoundError: If config file doesn't exist.
        yaml.YAMLError: If config file is invalid YAML.

    Example:
        >>> config = load_config("configs/negation_v0.yaml")
        >>> print(config["dataset"]["source_corpus"])
        'msmarco-passage'
    """
    # TODO: Implementation
    raise NotImplementedError("load_config not yet implemented")


def save_jsonl(data: list[dict], output_path: str | Path) -> None:
    """
    Save a list of dictionaries to a JSONL file.

    Args:
        data: List of dictionaries to save.
        output_path: Path to output JSONL file.

    Example:
        >>> examples = [{"id": "001", "query": "test"}]
        >>> save_jsonl(examples, "data/output.jsonl")
    """
    # TODO: Implementation
    raise NotImplementedError("save_jsonl not yet implemented")


def load_jsonl(input_path: str | Path) -> list[dict]:
    """
    Load a JSONL file into a list of dictionaries.

    Args:
        input_path: Path to input JSONL file.

    Returns:
        List of dictionaries, one per line.

    Raises:
        FileNotFoundError: If file doesn't exist.
        json.JSONDecodeError: If any line is invalid JSON.

    Example:
        >>> examples = load_jsonl("data/release/negation_v0/main.jsonl")
        >>> print(len(examples))
        2000
    """
    # TODO: Implementation
    raise NotImplementedError("load_jsonl not yet implemented")


def iter_jsonl(input_path: str | Path) -> Iterator[dict]:
    """
    Iterate over a JSONL file line by line (memory efficient).

    Args:
        input_path: Path to input JSONL file.

    Yields:
        Dictionary for each line.

    Example:
        >>> for example in iter_jsonl("data/large_file.jsonl"):
        ...     process(example)
    """
    # TODO: Implementation
    raise NotImplementedError("iter_jsonl not yet implemented")


def setup_logging(
    level: str = "INFO",
    log_file: str | Path | None = None
) -> logging.Logger:
    """
    Set up logging configuration.

    Args:
        level: Logging level (DEBUG, INFO, WARNING, ERROR).
        log_file: Optional path to log file.

    Returns:
        Configured logger instance.

    Example:
        >>> logger = setup_logging("DEBUG", "logs/run.log")
        >>> logger.info("Starting pipeline")
    """
    # TODO: Implementation
    raise NotImplementedError("setup_logging not yet implemented")


def set_seed(seed: int) -> None:
    """
    Set random seed for reproducibility.

    Sets seed for:
    - Python's random module
    - NumPy
    - PyTorch (if available)

    Args:
        seed: Random seed value.

    Example:
        >>> set_seed(42)
    """
    # TODO: Implementation
    raise NotImplementedError("set_seed not yet implemented")


def ensure_dir(path: str | Path) -> Path:
    """
    Ensure a directory exists, creating it if necessary.

    Args:
        path: Directory path.

    Returns:
        Path object for the directory.

    Example:
        >>> output_dir = ensure_dir("data/intermediate/candidates")
    """
    # TODO: Implementation
    raise NotImplementedError("ensure_dir not yet implemented")
